feat: refactor timezone service modules
