+++ src/user/services/user.service.ts
@@ -16,6 +16,7 @@ import { PageDto } from "../../common/dtos/page.dto";
 import { PageMetaDto } from "../../common/dtos/page-meta.dto";
 import { DbFilteredUser } from "../entities/filtered-users.entity";
 import { FilteredUsersDto } from "../dtos/filtered-users.dto";
+import { DbTimezone } from "../entities/timezones.entity";
 
 @Injectable()
 export class UserService {
@@ -325,5 +326,20 @@ export class UserService {
     }
 
     return item;
+  }
+
+  async getAllTimezones(): Promise<DbTimezone[]> {
+    const queryBuilder = this.baseQueryBuilder();
+
+    //  return all timezones that are not empty and not duplicates and not null
+    queryBuilder
+      .select("users.timezone as timezone")
+      .where("users.timezone IS NOT NULL")
+      .andWhere("users.timezone != ''")
+      .groupBy("users.timezone");
+
+    const timezones: DbTimezone[] = await queryBuilder.getRawMany();
+
+    return timezones;
   }
 }
